+++ Assets/Mvc/ControllerManager.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ModelViewController
+namespace Mvc
 {
     public static class ControllerManager
     {
