+++ Assets/Mvc/ICleareable.cs
@@ -1,4 +1,4 @@
-﻿namespace ModelViewController
+﻿namespace Mvc
 {
     public interface ICleareable
     {
