+++ Assets/Mvc/IMessageReceivable.cs
@@ -1,4 +1,4 @@
-﻿namespace ModelViewController
+﻿namespace Mvc
 {
     public interface IMessageReceivable
     {
